+++ gluecodium/src/test/resources/smoke/nesting/output/cpp/src/smoke/OuterStruct.cpp
@@ -3,7 +3,6 @@
 //
 // -------------------------------------------------------------------------------------------------
 #include "smoke/OuterStruct.h"
-#include <string>
 #include <utility>
 namespace smoke {
 OuterStruct::InnerStruct::InnerStruct( )
@@ -33,33 +32,5 @@ OuterStruct::OuterStruct( )
 OuterStruct::OuterStruct( ::std::string field )
     : field( std::move( field ) )
 {
-}
-std::error_code
-make_error_code( ::smoke::OuterStruct::InnerEnum value ) noexcept
-{
-    class InnerEnumErrorCategory: public ::std::error_category
-    {
-    public:
-        ~InnerEnumErrorCategory( ) override = default;
-        const char*
-        name( ) const noexcept override
-        {
-            return "InnerEnumErrorCategory";
-        }
-        std::string
-        message( int condition ) const override
-        {
-            switch( ::smoke::OuterStruct::InnerEnum( condition ) )
-            {
-            case( ::smoke::OuterStruct::InnerEnum::FOO ):
-                return "::smoke::OuterStruct::InnerEnum::FOO";
-            case( ::smoke::OuterStruct::InnerEnum::BAR ):
-                return "::smoke::OuterStruct::InnerEnum::BAR";
-            }
-            return "Unknown enum value";
-        }
-    };
-    static InnerEnumErrorCategory category{};
-    return std::error_code( static_cast<int>( value ), category );
 }
 }